Adding profiling.  To profile your program (or, rather, the underlying Python), pass "--profile" to adder.py.
The first pass, profiling samples/html.+, is not encouraging.  2/3 of
the time is spent in macroexpansion.  (Admittedly, html.+ is mostly
macros, but still.) And nearly all of the time spent in macroexpansion
is in Adder evaluation, which blows away the theory that the code runs
quickly once it's compiled.  Time to start on the optimizer? Or is
there some big mistake I've made in annotate_defmacro()?